+++ platforms/core-execution/persistent-cache/src/main/java/org/gradle/cache/FileLockManager.java
@@ -16,13 +16,13 @@
 package org.gradle.cache;
 
 import org.gradle.api.Action;
-import org.gradle.internal.service.scopes.Scope;
+import org.gradle.internal.service.scopes.Scopes;
 import org.gradle.internal.service.scopes.ServiceScope;
 
 import javax.annotation.Nullable;
 import java.io.File;
 
-@ServiceScope(Scope.Global.class)
+@ServiceScope(Scopes.Global.class)
 public interface FileLockManager {
     /**
      * Creates a lock for the given file with the given mode. Acquires a lock with the given mode, which is held until the lock is
