+++ platforms/core-runtime/base-annotations/src/main/java/org/gradle/internal/service/scopes/Scope.java
@@ -16,11 +16,8 @@
 
 package org.gradle.internal.service.scopes;
 
+/**
+ * @see Scopes
+ */
 public interface Scope {
-    /**
-     * These services are reused across builds in the same process.
-     *
-     * <p>Global services are visible to all other services.</p>
-     */
-    interface Global extends Scope {}
 }
