+++ platforms/core-runtime/build-operations/src/main/java/org/gradle/internal/operations/BuildOperationAncestryTracker.java
@@ -16,7 +16,7 @@
 
 package org.gradle.internal.operations;
 
-import org.gradle.internal.service.scopes.Scope;
+import org.gradle.internal.service.scopes.Scopes;
 import org.gradle.internal.service.scopes.ServiceScope;
 
 import javax.annotation.Nullable;
@@ -25,7 +25,7 @@ import java.util.function.Function;
 import java.util.function.Predicate;
 
 @SuppressWarnings("Since15")
-@ServiceScope(Scope.Global.class)
+@ServiceScope(Scopes.Global.class)
 public interface BuildOperationAncestryTracker {
     Optional<OperationIdentifier> findClosestMatchingAncestor(@Nullable OperationIdentifier id, Predicate<? super OperationIdentifier> predicate);
