+++ platforms/core-runtime/files/src/main/java/org/gradle/internal/file/Deleter.java
@@ -16,7 +16,7 @@
 
 package org.gradle.internal.file;
 
-import org.gradle.internal.service.scopes.Scope;
+import org.gradle.internal.service.scopes.Scopes;
 import org.gradle.internal.service.scopes.ServiceScope;
 
 import java.io.File;
@@ -25,7 +25,7 @@ import java.io.IOException;
 /**
  * A file deleter that doesn't give up if deletion doesn't work on the first try.
  */
-@ServiceScope(Scope.Global.class)
+@ServiceScope(Scopes.Global.class)
 public interface Deleter {
     /**
      * Attempts to delete the given file or directory recursively.
