+++ platforms/core-runtime/launcher/src/main/java/org/gradle/launcher/daemon/client/DaemonStartListener.java
@@ -17,7 +17,7 @@
 package org.gradle.launcher.daemon.client;
 
 import org.gradle.internal.service.scopes.EventScope;
-import org.gradle.internal.service.scopes.Scope.Global;
+import org.gradle.internal.service.scopes.Scopes.Global;
 import org.gradle.launcher.daemon.context.DaemonConnectDetails;
 
 /**
