+++ platforms/core-runtime/logging/src/main/java/org/gradle/internal/logging/progress/ProgressListener.java
@@ -20,7 +20,7 @@ import org.gradle.internal.logging.events.ProgressCompleteEvent;
 import org.gradle.internal.logging.events.ProgressEvent;
 import org.gradle.internal.logging.events.ProgressStartEvent;
 import org.gradle.internal.service.scopes.EventScope;
-import org.gradle.internal.service.scopes.Scope.Global;
+import org.gradle.internal.service.scopes.Scopes.Global;
 
 @EventScope(Global.class)
 public interface ProgressListener {
