+++ platforms/core-runtime/messaging/src/main/java/org/gradle/internal/event/ListenerManager.java
@@ -16,7 +16,7 @@
 
 package org.gradle.internal.event;
 
-import org.gradle.internal.service.scopes.Scope;
+import org.gradle.internal.service.scopes.Scopes;
 import org.gradle.internal.service.scopes.ServiceScope;
 
 /**
@@ -28,7 +28,7 @@ import org.gradle.internal.service.scopes.ServiceScope;
  * <p>Implementations are thread-safe: A listener is notified by at most 1 thread at a time, and so do not need to be thread-safe. All listeners
  * of a given type receive events in the same order. Listeners can be added and removed at any time.
  */
-@ServiceScope(Scope.Global.class)
+@ServiceScope(Scopes.Global.class)
 public interface ListenerManager {
     /**
      * Adds a listener.  A single object can implement multiple interfaces, and all interfaces are registered by a
