+++ platforms/core-runtime/process-services/src/main/java/org/gradle/process/internal/health/memory/JvmMemoryStatusListener.java
@@ -17,7 +17,7 @@
 package org.gradle.process.internal.health.memory;
 
 import org.gradle.internal.service.scopes.EventScope;
-import org.gradle.internal.service.scopes.Scope.Global;
+import org.gradle.internal.service.scopes.Scopes.Global;
 
 @EventScope(Global.class)
 public interface JvmMemoryStatusListener {
