+++ subprojects/core/src/main/java/org/gradle/api/internal/tasks/userinput/UserInputReader.java
@@ -16,13 +16,13 @@
 
 package org.gradle.api.internal.tasks.userinput;
 
-import org.gradle.internal.service.scopes.Scope;
+import org.gradle.internal.service.scopes.Scopes;
 import org.gradle.internal.service.scopes.ServiceScope;
 
 /**
  * Receives interactive responses from the user.
  */
-@ServiceScope(Scope.Global.class)
+@ServiceScope(Scopes.Global.class)
 public interface UserInputReader {
     void startInput();
