+++ subprojects/core/src/main/java/org/gradle/process/internal/ExecActionFactory.java
@@ -16,10 +16,10 @@
 
 package org.gradle.process.internal;
 
-import org.gradle.internal.service.scopes.Scope;
+import org.gradle.internal.service.scopes.Scopes;
 import org.gradle.internal.service.scopes.ServiceScope;
 
-@ServiceScope(Scope.Global.class)
+@ServiceScope(Scopes.Global.class)
 public interface ExecActionFactory {
     /**
      * Creates an {@link ExecAction} that is not decorated. Use this when the action is not made visible to the DSL.
